backend/requirements.txt
# backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.0
numpy==1.24.3
pandas==2.1.3
websockets==12.0
python-multipart==0.0.6
aiosqlite==0.19.0
httpx==0.25.2
apscheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4